oh there is a server station ai system
